Move Paths object to namer-core
