+++ namerd/iface/control-http/src/main/scala/io/buoyant/namerd/iface/AddrHandler.scala
@@ -6,7 +6,8 @@ import com.twitter.io.Buf
 import com.twitter.util.Future
 import io.buoyant.admin.names.DelegateApiHandler
 import io.buoyant.admin.names.DelegateApiHandler.{Addr => JsonAddr}
-import io.buoyant.namerd.{Paths, RichActivity}
+import io.buoyant.namer.Paths
+import io.buoyant.namerd.RichActivity
 
 object AddrUri extends NsPathUri {
   import HttpControlService._
